requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
qdrant-client>=1.7.0
openai>=1.3.0
groq>=0.4.0
python-decouple>=3.8
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
python-dotenv>=1.0.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
faker>=19.0.0

